Fixed code review comments and Unit tests.
